3rd Update index.html
Prompt to Gemini (forgot to mention pro edition):

yea this is great! Can you double check that the math with the customers and ingredients is working correctly? I think we can double check this too with more verbose JSON logs. The time stamp should be MM/DD/YYYY HH:MM:SS but for now we will just count hours 8:00 to 22:00 since that is when the restaurant will be open. The logs should show in UTC time or the timezone configured in the user's device, give the option in the logs and the UI of the game.