improved summary section with AP used previous day
prompt:

let's show a summary section at the top, we can just add to where the other game stats like day and timezone are. The amount of AP used the previous day, how many of the orders were deliveries, and the total number of customers served should be added. This will be helpful data for the user to plan staffing that day. In the logs it should show the amount of AP used that day